Adds feedback form
We want to encourage people to give us feedback to improve the form.

Every page now has a link to the feedback form on at the bottom.
This opens a Google Form, which stores the feedback in a Google
Spreadsheet. It doesn't end there though: when a new row is added, this
triggers creation of a GitHub Issue. Yay \o/